Further improvement in scripts/testall and make checkall to make them compatible
git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@1344 2bfe0521-f11c-4a00-b80e-6202646ff360

+++ scripts/test1.sh
@@ -1,7 +1,26 @@
 #! /bin/bash
 
 # Test a package
+# An option "--csl" or "--psl" can specify that only that one Lisp
+# is to be used.
 
+#    scripts/test1.sh [--csl or --psl] package_name
+
+
+csl="yes"
+psl="yes"
+case $1 in
+--csl)
+  psl="no"
+  shift
+  ;;
+--psl)
+  csl="no"
+  shift
+  ;;
+*)
+  ;;
+esac
 
 # I want this script to be one I can launch from anywhere, so
 # to access files etc I need to know where it lives.
@@ -41,8 +60,9 @@ here=`echo $here | sed -e 's+/[^/]*$++'`
 
 echo $here
 
-
+# If no argument is provided then this runs alg.tst
 p=${1:-alg}
+# WARNING - the "-n" option to echo is not portable.
 echo -n "Package to test is $p: "
 
 w=`grep " test " $here/packages/package.map | grep "($p "`
@@ -63,8 +83,17 @@ esac
 
 # This is bash-specific. If anybody really wanted this could
 # be redone using sed then this script could use plain /bin/sh.
+# It takes a line '(modname "dirname" other stuff)' and deletes up to and
+# including the first double quote and everything from the other double
+# quote onwards. What is left is the directory name within trunk/packages
+# that the test is in.
+
 d=${w%\"*}
 d=${d#*\"}
+
+# If I am running on Windows I need to have the file name in
+# (close to) native windows form. I can usefully retain "/" rather than
+# "\", but eg it will be in the form "d:/directory/file.ext"
 
 f="$here/packages/$d/$p.tst"
 if test -f /usr/bin/cygpath
@@ -72,43 +101,78 @@ then
   f=`cygpath -m $f`
 fi
 
+# I will annotate the strict with the identity of the machine on which the
+# test was run...
+
+mc=`$here/config.guess`
+mc=`$here/scripts/findhost.sh $mc`
+
+# There is one reall mess about the above identification of platform. On
+# a 64-bit windows the default behaviour is that a 64-bit version will be
+# built and used if possible. But possibility will depend on the availability
+# if build-tools etc. So I try tests here that will probably work...
+# In unusual cases it may mis-report but unless you are going to install
+# all the log files built here that is not very important!
+
+if test "x$mc" = "xi686-pc-windows" && test -d cslbuild/x86_64-w64-windows
+then
+  mc="x86_64-w64-windows"
+fi
+
+# Each individual test should only take a few seconds. On my computer the
+# slowest ones may take almost 15 seconds. The idea behind applying a
+# ulimit here is to avoid trouble when and if a test script loops.
+ 
+ulimit -c 60
+
+if test "$csl" = "yes"
+then
+
 mkdir -p csl-times
-
-ulimit -c 30
-
-# Note that resettime1/showtime1 is a new timing feature in Reduce
-# intended to support this and avoid any other timings taken within
-# test scripts from hurting.
 
 time ($here/bin/redcsl -w <<XXX > csl-times/$p.rlg.tmp) 2>howlong.tmp
 off int;
+symbolic linelength 80;
 symbolic(!*redeflg!* := nil);
 %off pwrds;
 on errcont;
 $loader
 resettime1;
-write "START OF REDUCE TEST RUN"$
+write "START OF REDUCE TEST RUN ON $mc"$
 in "$f";
 write "END OF REDUCE TEST RUN"$
-showtime1;
-quit;
+showtime1$
+quit$
 XXX
 cat howlong.tmp >> csl-times/$p.rlg.tmp
 echo -n CSL...
-sed -e "1d" -e "/^Time for test:/,//d" <$here/packages/$d/$p.rlg | sed -e '/^Total time taken:/d; /^Number of garbage/d; /^Time: /d; /^CRACK needed :/d; /^time for init/d; /^time to formulate/d' >csl-times/$p.rlg.orig
+sed -e "/^Tested on /,//d" <$here/packages/$d/$p.rlg | sed -e '/^Total time taken:/d; /^Number of garbage/d; /^Time: /d; /^CRACK needed :/d; /^time for init/d; /^time to formulate/d' >csl-times/$p.rlg.orig
 sed -e "1,/START OF REDUCE TEST RUN/d" -e "/END OF REDUCE TEST RUN/,//d" -e "/OMIT/,/TIMO/d" <csl-times/$p.rlg.tmp | sed -e "1s/^1: //" | sed -e '$s/^1: //' | sed -e '/^Total time taken:/d; /^Number of garbage/d; /^Time: /d; /^CRACK needed :/d; /^time for init/d; /^time to formulate/d' >csl-times/$p.rlg
 diff -B -w csl-times/$p.rlg.orig csl-times/$p.rlg >csl-times/$p.rlg.diff
-if test -s csl-times/$p.rlg.diff
-  then echo -n "Diff is in csl-times/$p.rlg.diff "
-  else echo -n "OK! " ; rm csl-times/$p.rlg.diff csl-times/$p.rlg.orig
+if test "$psl" = "yes"
+then
+  n="-n"
+else
+  n=""
 fi
-sed -e "1,/END OF REDUCE TEST RUN/d"  <csl-times/$p.rlg.tmp | sed -e '/^1: /d;' >csl-times/$p.time
+if test -s csl-times/$p.rlg.diff
+  then echo $n "Diff is in csl-times/$p.rlg.diff "
+  else echo $n "OK! " ; rm csl-times/$p.rlg.diff csl-times/$p.rlg.orig
+fi
+echo "Tested on $mc CSL" > csl-times/$p.time
+sed -e "1,/END OF REDUCE TEST RUN/d"  <csl-times/$p.rlg.tmp | sed -e '/^1: /d;' >>csl-times/$p.time
 rm csl-times/$p.rlg.tmp
+
+fi # CSL case
+
+if test "$psl" = "yes"
+then
 
 mkdir -p psl-times
 
 time ($here/bin/redpsl <<XXX > psl-times/$p.rlg.tmp) 2>howlong.tmp
 off int;
+symbolic linelength 80;
 symbolic(!*redefmsg := nil);
 symbolic(!*redeflg!* := nil);
 %off pwrds;
@@ -118,31 +182,34 @@ resettime1;
 write "START OF REDUCE TEST RUN"$
 in "$f";
 write "END OF REDUCE TEST RUN"$
-showtime1;
-quit;
+showtime1$
+quit$
 XXX
 cat howlong.tmp >> psl-times/$p.rlg.tmp
 echo -n "PSL..."
-sed -e "1d" -e "/^Time for test:/,//d" <$here/packages/$d/$p.rlg | sed -e '/^Total time taken:/d; /^Number of garbage/d; /^Time: /d; /^CRACK needed :/d; /^time for init/d; /^time to formulate/d' >psl-times/$p.rlg.orig
+sed -e "/^Tested on /,//d" <$here/packages/$d/$p.rlg | sed -e '/^Total time taken:/d; /^Number of garbage/d; /^Time: /d; /^CRACK needed :/d; /^time for init/d; /^time to formulate/d' >psl-times/$p.rlg.orig
 sed -e "1,/START OF REDUCE TEST RUN/d" -e "/END OF REDUCE TEST RUN/,//d" -e "/OMIT/,/TIMO/d" <psl-times/$p.rlg.tmp | sed -e "1s/^1: //" | sed -e '$s/^1: //' | sed -e '/^Total time taken:/d; /^Number of garbage/d; /^Time: /d; /^CRACK needed :/d; /^time for init/d; /^time to formulate/d' >psl-times/$p.rlg
 diff -B -w psl-times/$p.rlg.orig psl-times/$p.rlg >psl-times/$p.rlg.diff
 if test -s psl-times/$p.rlg.diff
   then echo "diff is in psl-times/$p.rlg.diff"
   else echo "OK! " ; rm psl-times/$p.rlg.diff psl-times/$p.rlg.orig
 fi
+echo "Tested on $mc PSL" > psl-times/$p.time
 sed -e "1,/END OF REDUCE TEST RUN/d"  <psl-times/$p.rlg.tmp | sed -e '/^1: /d;' >psl-times/$p.time
 rm  psl-times/$p.rlg.tmp
 
-mkdir -p csl-psl-times-comparison
-diff -B -w csl-times/$p.rlg psl-times/$p.rlg >csl-psl-times-comparison/$p.rlg.diff
-if test -s csl-psl-times-comparison/$p.rlg.diff
-  then echo "CSL and PSL test logs differ!"
-  else rm csl-psl-times-comparison/$p.rlg.diff
+fi # PSL case
+
+if test "$csl" = "yes" && test "$psl" = "yes"
+then
+  mkdir -p csl-psl-times-comparison
+  diff -B -w csl-times/$p.rlg psl-times/$p.rlg >csl-psl-times-comparison/$p.rlg.diff
+  if test -s csl-psl-times-comparison/$p.rlg.diff
+    then echo "CSL and PSL test logs differ!"
+    else rm csl-psl-times-comparison/$p.rlg.diff
+  fi
 fi
 
-rm howlong.tmp
+rm -f howlong.tmp
 
 # end of test
-
-
-
